refactor: chart
